requirements.txt
requests
python-dotenv
pandas
datetime
pandas
sklearn
sendgrid==6.0.5
seaborn
matplotlib
